Parses SDK version when creating the ngx system. If version is not supported by dlss returns error
