+++ crates/nvngx-sys/Cargo.toml
@@ -21,7 +21,6 @@ include = [
 [build-dependencies]
 bindgen = { version = "0.72", optional = true }
 cc = "1"
-regex = "1"
 
 [dependencies]
 widestring = "1"
