+++ crates/nvngx/Cargo.toml
@@ -17,6 +17,7 @@ log = "0.4"
 uuid = { version = "1", features = ["v4"] }
 derive_builder = "0.12"
 nvngx-sys = { version = "0.3.0", path = "../nvngx-sys" }
+regex = "1"
 
 [target.'cfg(windows)'.dependencies]
 windows = { version = "0.61", features = ["Win32_Graphics_Direct3D12", "Win32_Graphics_Dxgi_Common"], optional = true }
